+++ crates/noirc_abi/src/input_parser/toml.rs
@@ -1,4 +1,4 @@
-use super::InputValue;
+use super::{parse_str_to_field, InputValue};
 use crate::{errors::InputParserError, Abi, AbiType, MAIN_RETURN_NAME};
 use acvm::FieldElement;
 use iter_extended::{btree_map, try_btree_map, try_vecmap, vecmap};
@@ -151,33 +151,6 @@ impl InputValue {
         };
 
         Ok(input_value)
-    }
-
-    // TODO: move this into a different module that toml
-    pub fn try_from_cli_args(
-        value: String,
-        param_type: &AbiType,
-    ) -> Result<InputValue, InputParserError> {
-        let input_value = match param_type {
-            AbiType::String { .. } => InputValue::String(value),
-            AbiType::Field | AbiType::Integer { .. } => {
-                InputValue::Field(parse_str_to_field(&value)?)
-            }
-            _ => return Err(InputParserError::AbiTypeMismatch(param_type.clone())),
-        };
-        // TODO: support arrays
-        Ok(input_value)
-    }
-}
-
-fn parse_str_to_field(value: &str) -> Result<FieldElement, InputParserError> {
-    if value.starts_with("0x") {
-        FieldElement::from_hex(value).ok_or_else(|| InputParserError::ParseHexStr(value.to_owned()))
-    } else {
-        value
-            .parse::<i128>()
-            .map_err(|err_msg| InputParserError::ParseStr(err_msg.to_string()))
-            .map(FieldElement::from)
     }
 }
